+++ games/static/js/add_edition.js
@@ -13,6 +13,12 @@ let syncData = [
     "target": "#id_sort_name",
     "target_value": "value"
   },
+  {
+    "source": "#id_game",
+    "source_value": "dataset.year",
+    "target": "#id_year_released",
+    "target_value": "value"
+  },  
 ]
 
 syncSelectInputUntilChanged(syncData, "form");