+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+### Improved
+* game overview: simplify playtime range display
+
 ## 1.3.0 / 2023-11-05 15:09+01:00
 
 ### New