+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 
 ### Improved
 * game overview: simplify playtime range display
+* new session: order devices alphabetically
 
 ## 1.3.0 / 2023-11-05 15:09+01:00